+++ tests/error_origin.js
@@ -21,4 +21,13 @@ module.exports = async ({ test, assert, affirm }) => [
         const tr = await test("?>", async () => assert(false, true))
         affirm(tr.trace, (trace) => trace.includes('error_origin.js'))
     })
+    // // TODO for tests that cannot be validated within run.js scope:
+    // // create [fail tests] folder with expected failurures & call them via shell
+    // // Manual tests (uncomment to see errors):
+    // , (async () => test("Sync test without await gives correct error", () => {
+    //     throw 'err'
+    // }))()
+    // , (async () => test("Async test without await gives await hint on error", async () => {
+    //     throw 'err'
+    // }))()
 ]