+++ docs/CHANGELOG.md
@@ -4,6 +4,11 @@
 * **m**aintenance
 * **p**erformance
 
+### 2020.07.26
+
+ * **f** Add hint: missing await for async test
+ * **m** Remove emeraldwalk .vscode as default live test runner
+
 ### 2020.07.25
 
  * **f** Add hint: missing test function in test file.